parser objects

+++ srcs/init/init.c
@@ -21,9 +21,9 @@ t_cam	cam_init(t_scene *s)
 	cam->pos = set_vec(0, 0, 0);
 	cam->orientation = set_vec(0, 0, -1);
 	cam->fov_w = 70;
-	fov_w_half_len = tan(to_radian((float)cam->fov_w / 2));
+	fov_w_half_len = tanf(to_radian((float)cam->fov_w / 2));
 	cam->fov_h_len = 2 * fov_w_half_len / s->image_ratio;
-	cam->fov_h = (int)to_degree(atan(cam->fov_h_len / 2)) * 2;
+	cam->fov_h = (int)to_degree(atanf(cam->fov_h_len / 2)) * 2;
 	cam->translate = set_vec(0, 0, 0);
 	cam->rotate = set_vec(0, 0, 0);
 	// we would need this to do translation. codes are written but never tested yet.
